refactor to separate individual config observation contributors
